+++ Mailer/extensions/mailer-module/src/migrations/20240610073219_change_mail_templates_name_column_length.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+use Phinx\Migration\AbstractMigration;
+
+final class ChangeMailTemplatesNameColumnLength extends AbstractMigration
+{
+    public function up(): void
+    {
+        $this->table('mail_templates')
+            ->changeColumn('name', 'string', [
+                'length' => 768,
+                'null' => false,
+            ])
+            ->update();
+    }
+
+    public function down(): void
+    {
+        $this->output->writeln('Down migration is not available. Reverting the column length may cause data truncation.');
+    }
+}